Anime Update
Visually this is a much needed fix. When loading the addRecord page for an anime update this loads all the basic info first and then separately loads the related content. While the related content is loaded, the preloader is active to indicate to the user that something is still loading.

At the very least the most basic details are shown while the rest loads, which can take a while depending on the amount of related content available.