Small Improvements
Moving some variables to be required only when necessary. This should lighten the initial application load.

Also fixed issue where xlsx import process did not copy over assets properly if available. This arose from the fact that the xlsx and zip files are no longer named the same since the zip file generated is the same regardless of whether the xlsx import is chosen to be simple or detailed.