Index Page Issues
Fixes the issue where the vertical scroll bar would appear for the whole page.

Fixes the issue where anime records were not displaying the correct rating due to some of the elements being used in the calculation showing up as strings. The same issue has been resolved for show records as well.